+++ app/cells/folio/console/tagsinput_cell.rb
@@ -3,33 +3,40 @@
 class Folio::Console::TagsinputCell < Folio::ConsoleCell
   include SimpleForm::ActionViewExtensions::FormHelper
 
-  def separator
-    options[:separator] || ' '
+  def show
+    model.input name, input_html: input_html,
+                      wrapper_html: wrapper_html
   end
 
-  def value
-    options[:value] || model.object.tag_list.join(separator)
+  def selected
+    options[:selected] || model.object.tag_list
   end
 
   def name
     options[:name] || :tag_list
   end
 
-  def values
-    options[:values] || ActsAsTaggableOn::Tag.limit(1000).all.map(&:name)
+  def collection
+    options[:collection] || ActsAsTaggableOn::Tag.pluck(:name)
   end
 
   def allow_creation
-    !options[:disable_creation]
+    options[:disable_creation] ? nil : true
   end
 
   def input_html
+    value = model.object.send(name).presence
+    value = value.join(', ') if value.is_a?(Array)
     {
-      class: 'folio-tagsinput',
-      'data-tags': values.join(separator),
-      'data-allow-create': allow_creation,
-      'data-comma-separated': (/,\s*/.match?(separator) ? 'true' : nil),
+      class: 'folio-console-tagsinput',
       value: value,
+    }
+  end
+
+  def wrapper_html
+    {
+      'data-allow-create': allow_creation,
+      'data-collection': collection.join(', '),
     }
   end
 end